+++ src/docs/views/components/form/page-form-checkbox-group/page-form-checkbox-group.vue
@@ -28,12 +28,44 @@
 				<ul>
 					<li>An array of strings <code>[&quot;option1&quot;, &quot;option2&quot;, &quot;option3&quot;]</code>, in which case both the label and value are the same.</li>
 					<li>An array of numbers <code>[1, 2, 3]</code>, in which case both the label and value are the same.</li>
-					<li>An object <code>{ value: &quot;label&quot; }</code> where each key is the value of an option, and each value is the label.</li>
 					<li>An array of objects <code>[{ label: &quot;Label&quot;, value: &quot;value&quot; }]</code> where labels and values are explicitly defined.</li>
+					<li>An array of objects combined with the <code>labelKey</code> and <code>valueKey</code> properties.</li>
+					<li>An object <code>{ value: &quot;label&quot; }</code> where each key is the value of oan option, and each value is the label.</li>
 				</ul>
-
-				<p>It is possible to combine both array options, but generally not recommended.</p>
 			</component-prop>
+
+			<component-prop id="prop-label-key">
+				<template #name>
+					labelKey
+				</template>
+
+				<template #type>
+					String
+				</template>
+
+				<template #default-value>
+					label
+				</template>
+
+				The key needed to find each option's label within its object. If an individual option is a string or number, this is ignored.
+			</component-prop>
+
+			<component-prop id="prop-value-key">
+				<template #name>
+					valueKey
+				</template>
+
+				<template #type>
+					String
+				</template>
+
+				<template #default-value>
+					value
+				</template>
+
+				The key needed to find each option's value within its object. If an individual option is a string or number, this is ignored.
+			</component-prop>
+
 			<component-prop id="prop-name">
 				<template #name>
 					name
@@ -49,6 +81,7 @@
 
 				<p>A name for this checkbox group. If not set, the input ID is used.</p>
 			</component-prop>
+
 			<component-prop id="prop-id">
 				<template #name>
 					id
@@ -75,6 +108,7 @@
 				<p>The default slot contains the label of the field.</p>
 				<p><em>Note that a label is always required, even if a label is not shown to the user, as it will be critical for screen readers. This component will show an error if no label is provided.</em></p>
 			</component-slot>
+
 			<component-slot id="slot-introduction">
 				<template #name>
 					introduction
@@ -82,6 +116,7 @@
 
 				<p>Any additional text to introduce the options, which appears between the label and options.</p>
 			</component-slot>
+
 			<component-slot id="slot-error">
 				<template #name>
 					error
@@ -89,6 +124,7 @@
 
 				<p>Any error text to display below the field.</p>
 			</component-slot>
+
 			<component-slot id="slot-help">
 				<template #name>
 					help