+++ support/templates/composable/composable.test.js
@@ -1,14 +1,14 @@
 import { describe, expect, test } from "vitest";
-import {{CAMEL_CASE_NAME}} from "./{{COMPOSABLE_NAME}}.vue";
+import {{CAMEL_CASE_NAME}} from "./{{COMPOSABLE_NAME}}";
 
 describe("{{COMPOSABLE_NAME}}", () => {
-	test("should initialise", () => {
+	test("The composable initialises", () => {
 		const response = {{CAMEL_CASE_NAME}}();
 
 		expect(response).toBeTypeOf("object");
 	});
 
-	test("should ", () => {
+	test("", () => {
 		const { returned } = {{CAMEL_CASE_NAME}}();
 	});
 });